a_preproc_SMI_ETG: gaze data now synced to video.
SMI appears to double up (or more) some frames from the scene camera to keep gaze data in sync with it. The logica by which it does so is unknown, nor is there anything available in their data export that it seems to be based on. The solution is to use Begaze to export the scene video with an invisible gaze overlay and use that instead of the raw avi and wav files. Then still a bit of empirical fiddling was needed to get things to line up..